web/flaskproject/doctorDecorator.py
from flask import request, session
from flaskproject import app
from flask.json import jsonify
from functools import wraps
import jwt


def doctoken_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = session.get('docjwt') 
        # return 401 if token is not passed
        if not token:
            return jsonify({'message' : 'Token is missing !!'}), 401
  
        try:
            # decoding the payload to fetch the stored details
            print(token)
            data = jwt.decode(token, app.config['SECRET_KEY'],algorithms=["HS256"])
            print(data)
            current_user = data['user']

        except:
            return jsonify({
                'message' : 'Token is invalid !!'
            }), 401
        # returns the current logged in users contex to the routes
        return  f(current_user, *args, **kwargs)
  
    return decorated