web/requirements.txt

  
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
pyJWT
gunicorn
pytest